backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
Pillow==10.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7

# Celery and Redis for background tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# External API clients
requests==2.31.0

# SMS/USSD integration
africastalking==1.2.6
